refactor: carry cdk sources into separate folder (#14)
