+++ cdk/cdk-workshop-stack.ts
@@ -1,4 +1,4 @@
-import { Stack, StackProps } from 'aws-cdk-lib'
+import { Stack, type StackProps } from 'aws-cdk-lib'
 import { LambdaRestApi } from 'aws-cdk-lib/aws-apigateway'
 import { Construct } from 'constructs'
 import { HitCounter } from './constructs/hitcounter.js'
