+++ tsconfig.vitest.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.app.json",
-  "include": ["src", "tests"],
+  "include": ["cdk", "src", "tests"],
   "exclude": [],
   "compilerOptions": {
     "composite": true,
